rebased and added reusable view container
